SEXPR SymbolicExpression(R)
vs. Symbolic(Expression(R)), i.e. everything symbolic is an Expresson.

Expressions involving symbolic functions that remain unevaluated and
are displayed as entered. Evaluation is delayed until an operation is
performed which requires a canonical form such as testing for equality or
simplification. The domain depends directly internally on the Expression
domain for evaluation of expressions.

This patch also includes minor changes to float and fraction to cleanup the
conversion to InputForm.